delete Reserved
