+++ Reservation.py
@@ -1,8 +1,10 @@
-class Reservation:
-    def __init__(self,check_in,reservation_detail,calendar):
-        self._check_in = check_in
-        self._reservation = reservation_detail
-        self._calendar = calendar
+from Room import Room
 
-    def get_detail_payment(self):
+class Reservation(Room):
+    def __init__(self,check_in,reservation_detail,calendar,room_id,room_rental,room_status,room_fac):
+        Room.__init__(self,room_id,room_rental,room_status,room_fac)
+        self.__check_in = check_in
+        self.__reservation_detail = reservation_detail
+        self.__calendar = calendar
+    def get_detail_payment():
         pass
